[cli/plugin] Removed open options to cleanly overwrite existing files
Previously, if the target files already existed, the exisiting file was
not purged before the code was written. This could lead to remnants of a
previous generation being left over if the new code was shorter than
before.

The default nio.write() options are CREATE, WRITE and TRUNCATE_EXISTING,
which implement exactly the desired behaviour.

Signed-off-by: Nis Wechselberg <wechselberg.nisboerge@scheidt-bachmann-st.de>